+++ retz-server/src/main/java/io/github/retz/scheduler/JobQueue.java
@@ -19,10 +19,12 @@ package io.github.retz.scheduler;
 import io.github.retz.cli.TimestampHelper;
 import io.github.retz.protocol.StatusResponse;
 import io.github.retz.protocol.data.Job;
+import org.apache.mesos.Protos;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
+import java.sql.Time;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Optional;
@@ -199,5 +201,49 @@ public class JobQueue {
     // Methods for test
     public static void setStatus(StatusResponse response) {
         response.setStatus(JobQueue.size(), countRunning());
+    }
+
+    /*
+     * Update local job state according to status in Mesos
+     *
+                      | STARTING   | STARTED
+        --------------+------------+------------
+        TASK_RUNNING  | STARTED    | keep
+        TASK_STAGING  | keep       | -
+        TASK_STARTING | keep       | -
+        TASK_LOST     | QUEUED     | QUEUED
+        TASK_FINISHED | FINISHED   | FINISHED
+        TASK_KILLED   | KILLED     | KILLED
+        TASK_ERROR    | KILLED     | KILLED
+     */
+    public static Job updateJobStatus(Job job, String stateInMesos) {
+        Protos.TaskState taskState = Protos.TaskState.valueOf(stateInMesos);
+        switch (taskState.getNumber()) {
+            case Protos.TaskState.TASK_STAGING_VALUE:
+            case Protos.TaskState.TASK_STARTING_VALUE:
+                break;
+            case Protos.TaskState.TASK_RUNNING_VALUE:
+                if (job.state() == Job.JobState.STARTING) {
+                    job.started(job.taskId(), Optional.empty(), TimestampHelper.now());
+                }
+                break;
+            case Protos.TaskState.TASK_LOST_VALUE:
+                job.doRetry();
+                break;
+            case Protos.TaskState.TASK_FINISHED_VALUE:
+                //TODO: recover all information right here
+                job.finished(TimestampHelper.now(), Optional.empty(), 0);
+                break;
+            case Protos.TaskState.TASK_KILLED_VALUE:
+            case Protos.TaskState.TASK_KILLING_VALUE:
+                job.killed(TimestampHelper.now(), Optional.empty(), "KILLED");
+                break;
+            case Protos.TaskState.TASK_ERROR_VALUE:
+                job.killed(TimestampHelper.now(), Optional.empty(), "ERROR");
+                break;
+            default:
+                LOG.error("Unknown state: {}", stateInMesos);
+        }
+        return job;
     }
 }