+++ retz-server/src/test/java/io/github/retz/scheduler/MesosHTTPFetcherTest.java
@@ -16,10 +16,13 @@
  */
 package io.github.retz.scheduler;
 
+import org.apache.mesos.Protos;
 import org.junit.Test;
 
 import java.io.IOException;
 import java.io.InputStream;
+import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 
 import static org.hamcrest.MatcherAssert.assertThat;
@@ -55,5 +58,14 @@ public class MesosHTTPFetcherTest {
         assertTrue(s.isPresent());
         assertThat(s.get(),
                 is("/tmp/mesos/slaves/6c751ae7-6856-4127-aea1-42f3a9210846-S0/frameworks/3a3e9491-84a5-4c9d-8fed-5ca10c23d922-0000/executors/sum/runs/927b4c8a-bcfb-40fb-bf24-fcd4a430e2aa"));
+    }
+
+    @Test
+    public void parseTask() throws IOException {
+        InputStream in = MesosHTTPFetcherTest.class.getResourceAsStream("/master-tasks.json");
+        List<Map<String, Object>> result = MesosHTTPFetcher.parseTasks(in, "bcc4637f-34d7-4cab-8429-7146fabd7198-0001");
+        System.err.println(result);
+        assertThat(result.size(), is(1));
+        assertThat(result.get(0).get("state"), is("TASK_FINISHED"));
     }
 }